[JENKINS-48946] - Move whitelist ordering test to core to fail fast
